requirementstxt.txt

### **2. requirements.txt** (Dependencies)
```txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.22.0
transformers>=4.35.0
gradio>=4.12.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
imageio>=2.31.0
imageio[ffmpeg]
moviepy>=1.0.3
accelerate>=0.24.0
safetensors>=0.4.0
xformers>=0.0.22